test(analysis): separate benchmark tooling coverage
